move module source parsing to hcl2cdk to reuse code
