+++ packages/@cdktf/hcl2cdk/lib/terraform-dependency-constraint.ts
@@ -1,0 +1,7 @@
+export interface TerraformDependencyConstraint {
+  readonly name: string;
+  readonly source: string;
+  readonly version?: string;
+  readonly fqn: string;
+  readonly namespace?: string;
+}
