+++ packages/cdktf-cli/test/get/generator/module-generator.test.ts
@@ -2,7 +2,7 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import { ConstructsMaker, Language } from "../../../lib/get/constructs-maker";
-import { TerraformModuleConstraint } from "../../../lib/config";
+import { TerraformModuleConstraint } from "@cdktf/hcl2cdk";
 import { expectModuleToMatchSnapshot } from "../util";
 
 test("generate some modules", async () => {
